requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
tabulate==0.9.0
python-dotenv
